+++ client/src/App.jsx
@@ -3,33 +3,10 @@ import reactLogo from "./assets/react.svg";
 import viteLogo from "/vite.svg";
 import "./App.css";
 import EmbeddedPlayer from "./components/EmbeddedPlayer";
-import FormTextInput from "./components/FormTextInput";
+import AddSongForm from "./components/AddSongForm";
 
 function App() {
   const [count, setCount] = useState(0);
-  const [formData, setFormData] = useState({});
-  const server_url = import.meta.env.VITE_SERVER_URL;
-
-  function handleChangeEventFormInput(key, value) {
-    setFormData((prevFormData) => {
-      return {
-        ...prevFormData,
-        [key]: value
-      };
-    })
-  }
-
-  async function submitFormData() {
-    const sendSongsToDB = await fetch(`${server_url}/songs/add`, {
-      method: "POST",
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(formData)
-    })
-    const response = await sendSongsToDB.json()
-    alert(response.result)
-  }
   
   return (
     <>
@@ -42,10 +19,7 @@ function App() {
         </a>
       </div>
       <EmbeddedPlayer />
-      <FormTextInput inputName="Artist" inputID="artist" handleChangeEventFormInput={handleChangeEventFormInput}/>
-      <FormTextInput inputName="Song Title" inputID="song_title" handleChangeEventFormInput={handleChangeEventFormInput}/>
-      <FormTextInput inputName="Spotify URL" inputID="spotify_url" handleChangeEventFormInput={handleChangeEventFormInput}/>
-      <button onClick={submitFormData}>Submit</button>
+      <AddSongForm />
       <h1>Vite + React</h1>
       <div className="card">
         <button onClick={() => setCount((count) => count + 1)}>
